% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProjectLiver.R
\docType{class}
\name{TrenaProjectLiver-class}
\alias{TrenaProjectLiver-class}
\alias{.TrenaProjectLiver}
\alias{TrenaProjectLiver}
\title{TrenaProjectLiver-class}
\usage{
TrenaProjectLiver(quiet = TRUE)
}
\value{
An object of the TrenaProjectLiver class
}
\description{
Expression, variant and covariate data for the genes of interest (perhaps unbounded) for pre-term birth studies
}
